requirements.txt
Django==2.1
djangorestframework==3.8.2
django-rest-multiple-models
django-filter
Pillow